rename and cleanup more
